Piece Positioning and Table Dimensioning
So. This was a little tougher to crack. Yes. The Create() function goes through every single piece and puts it in the right place. But. The pieces are not aligned... it was not fun refactoring all those positions. but i got a function going. it looks exactly as it should now